requirements.txt
numpy
scipy
pandas
torch
torchvision
opencv-python
nuscenes-devkit
pyquaternion
pyyaml
tqdm
transformers
accelerate
open3d
Pillow
timm
